fix: Reorder flags in command generators
I've updated the `dbt_run.py` and `dbt_test.py` scripts to move the
`*cmd_flag` lists to be added to the `cmd` list prior to the model
section. This will prevent misclassification of cmd_flags as being
models by dbt in the event that we provide a --project-dir flag.